[AWIBOF-3870] fix type error
Former-commit-id: f1d766b523ff06ca7f3cf6026b55402e2b32089e

+++ src/telemetry/telemetry_client/telemetry_data_pool.cpp
@@ -62,7 +62,7 @@ TelemetryDataPool::SetLog(std::string id, uint32_t value)
     auto entry = pool.find(id);
     if (entry == pool.end())
     {
-        if (pool.size() == maxEntry)
+        if ((int)pool.size() == maxEntry)
         {
             POS_TRACE_ERROR(EID(TELEMETRY_CLIENT_ERROR), "[Telemetry] error!! failed to add new entry, num of entries reached max limit. Item:{}", id);
             return;
